Added spec_utils tests.
